Move and rename test
